Rehearsal page is now a server component
